refactor preview queries
