--- src/js/queries/getDepartmentPageRevisionQuery.js
@@ -1,15 +1,0 @@
-import departmentPageFragment from './departmentPageFragment';
-
-const getDepartmentPageRevisionQuery = `
-  query getPageRevision($id: ID) {
-    pageRevision(id: $id) {
-      id
-      asDepartmentPage {
-        ...departmentPageInfo
-      }
-    }
-  }
-  ${departmentPageFragment}
-`;
-
-export default getDepartmentPageRevisionQuery;
