--- src/js/queries/getEventPageRevisionQuery.js
@@ -1,15 +1,0 @@
-import eventPageFragment from './eventPageFragment';
-
-const getEventPageRevisionQuery = `
-  query getPageRevision($id: ID) {
-    pageRevision(id: $id) {
-      id
-      asEventPage {
-        ...eventPageInfo
-      }
-    }
-  }
-  ${eventPageFragment}
-`;
-
-export default getEventPageRevisionQuery;
