--- src/js/queries/getLocationPageRevisionQuery.js
@@ -1,15 +1,0 @@
-import locationPageFragment from './locationPageFragment';
-
-const getLocationPageRevisionQuery = `
-  query getPageRevision($id: ID) {
-    pageRevision(id: $id) {
-      id
-      asLocationPage {
-        ...locationPageInfo
-      }
-    }
-  }
-  ${locationPageFragment}
-`;
-
-export default getLocationPageRevisionQuery;
