--- src/js/queries/getOfficialDocumentPageRevisionQuery.js
@@ -1,15 +1,0 @@
-import officialDocumentPageFragment from './officialDocumentPageFragment';
-
-const getOfficialDocumentPageRevisionQuery = `
-  query getPageRevision($id: ID) {
-    pageRevision(id: $id) {
-      id
-      asOfficialDocumentPage {
-        ...officialDocumentPageInfo
-      }
-    }
-  }
-  ${officialDocumentPageFragment}
-`;
-
-export default getOfficialDocumentPageRevisionQuery;
